chore: add initializer file
