+++ cmd/mojave/lib.rs
@@ -1,5 +1,6 @@
 pub mod cli;
 pub mod command;
+pub mod initializer;
 pub mod logging;
 pub mod networks;
 pub mod options;
