--- crates/full-node/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "mojave-chain-full-node"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-mojave-chain-json-rpc = { workspace = true }
-
-# error handling
-thiserror = { workspace = true }
