--- crates/full-node/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod dummy;
