--- crates/json-rpc/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "mojave-chain-json-rpc"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-hyper = "1.6.0"
-jsonrpsee = { version = "0.24", features = ["server"] }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tower = "0.4"
-tower-http = { version = "0.6", features = ["cors"] }
-
-# misc
-tracing = { workspace = true }
-trait-variant = { workspace = true }
