--- crates/json-rpc/src/api/eth_pubsub.rs
@@ -1,11 +1,0 @@
-#[trait_variant::make(EthPubSubApi: Send)]
-pub trait LocalEthPubSubApi: Clone + Send + Sync + 'static {
-    // async fn subscribe_new_heads(&self) -> impl Stream<Item = Header<AnyHeader>> + Send + Unpin;
-
-    // async fn subscribe_logs(
-    //     &self,
-    //     filter: Option<Box<Filter>>,
-    // ) -> impl Stream<Item = Log> + Send + Unpin;
-
-    // async fn subscribe_new_pending_transaction(&self) -> impl Stream<Item = B256> + Send + Unpin;
-}
