--- crates/json-rpc/src/api/mod.rs
@@ -1,4 +1,0 @@
-pub mod eth;
-pub mod eth_pubsub;
-pub mod net;
-pub mod web3;
