--- crates/json-rpc/src/filter.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, Clone, Default)]
-pub struct EthFilter {}
-
-impl EthFilter {}
-
-pub struct Filter {}
