--- crates/json-rpc/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod api;
-pub mod config;
-pub mod error;
-pub mod filter;
-pub mod server;
